More progress on dsFAT

+++ include/Defs.h
@@ -4,7 +4,7 @@
 #include <cstdint>
 
 using    fd_t = uint64_t;
-using  mode_t = unsigned short;
+using  mode_t = unsigned int;
 using   dev_t = uint32_t;
 using inode_t = size_t;
 using   uid_t = unsigned int;
